added registration/login/signout functionality
Main changes are going to be in register.php/html and login.php/html. You will notice in both documents the username, password and dbname are blank. For testing you will have to first make sure you have the user Table set up in mariadb. You will then give register and login .php your mariadb user and password so it can access the database (eventually we will have to just use one of our databases so everything stored is synced). But then when using those files you should be able to register/login/signout test users. Owners are hardcoded in.

the table i have in mariadb is named 'users' and has these attributes:

+-----------+-------------------------------------+------+-----+---------+----------------+
| Field     | Type                                | Null | Key | Default | Extra          |
+-----------+-------------------------------------+------+-----+---------+----------------+
| userID    | int(11)                             | NO   | PRI | NULL    | auto_increment |
| username  | varchar(255)                        | NO   | UNI | NULL    |                |
| password  | varchar(255)                        | NO   |     | NULL    |                |
| user_type | enum('employee','customer','owner') | NO   |     | NULL    |                |
| email     | varchar(100)                        | NO   |     | NULL    |                |
| phone     | varchar(20)                         | YES  |     | NULL    |                |
+-----------+-------------------------------------+------+-----+---------+----------------+

Once a user is signed in they will be relocated to a homepage depending on their type with their corresponding links.

Additionally changed index from a .php to a .html since its not using any php and we are separating html and php files anyway.

Removed the duplicate product pdf.